@@ -1,22 +1,6 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\Mapping;
 
@@ -24,24 +8,15 @@ use Doctrine\Common\Persistence\Mapping\ClassMetadataFactory as BaseClassMetadat
 
 /**
  * Class responsible for retrieving ClassMetadata instances.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 class ClassMetadataFactory implements BaseClassMetadataFactory
 {
-    /**
-     * @var ClassMetadataInstantiatorInterface
-     */
+    /** @var ClassMetadataInstantiatorInterface */
     private $classMetadataInstantiator;
 
-    /**
-     * @var array
-     */
-    protected $classes = array();
+    /** @var ClassMetadataInterface[] */
+    private $classes = [];
 
-    /**
-     * @param ClassMetadataInstantiatorInterface $classMetadataInstantiator
-     */
     public function __construct(ClassMetadataInstantiatorInterface $classMetadataInstantiator)
     {
         $this->classMetadataInstantiator = $classMetadataInstantiator;
@@ -50,23 +25,19 @@ class ClassMetadataFactory implements BaseClassMetadataFactory
     /**
      * Returns all mapped classes.
      *
-     * @return array The ClassMetadata instances of all mapped classes.
+     * @return ClassMetadataInterface[] The ClassMetadataInterface instances of all mapped classes.
      */
-    public function getAllMetadata()
+    public function getAllMetadata() : array
     {
         return $this->classes;
     }
 
     /**
-     * Gets the class metadata descriptor for a class.
-     *
-     * @param string $className The name of the class.
-     *
-     * @return ClassMetadata
+     * @param string $className
      */
-    public function getMetadataFor($className)
+    public function getMetadataFor($className) : ClassMetadataInterface
     {
-        if (!isset($this->classes[$className])) {
+        if (! isset($this->classes[$className])) {
             $metadata = $this->classMetadataInstantiator->instantiate($className);
 
             if ($metadata->reflClass->implementsInterface('Doctrine\SkeletonMapper\Mapping\LoadMetadataInterface')) {
@@ -80,22 +51,16 @@ class ClassMetadataFactory implements BaseClassMetadataFactory
     }
 
     /**
-     * Checks whether the factory has the metadata for a class loaded already.
-     *
-     * @param string $className
-     *
-     * @return bool TRUE if the metadata of the class in question is already loaded, FALSE otherwise.
+     * {@inheritDoc}
      */
-    public function hasMetadataFor($className)
+    public function hasMetadataFor($className) : bool
     {
         return isset($this->classes[$className]);
     }
 
     /**
-     * Sets the metadata descriptor for a specific class.
-     *
-     * @param string        $className
-     * @param ClassMetadata $class
+     * @param string                 $className
+     * @param ClassMetadataInterface $class
      */
     public function setMetadataFor($className, $class)
     {
@@ -103,14 +68,9 @@ class ClassMetadataFactory implements BaseClassMetadataFactory
     }
 
     /**
-     * Returns whether the class with the specified name should have its metadata loaded.
-     * This is only the case if it is either mapped directly or as a MappedSuperclass.
-     *
-     * @param string $className
-     *
-     * @return bool
+     * {@inheritDoc}
      */
-    public function isTransient($className)
+    public function isTransient($className) : bool
     {
         return isset($this->classes[$className]);
     }
