@@ -1,33 +1,20 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\ObjectRepository;
+
+use BadMethodCallException;
+use function get_class;
 
 class BasicObjectRepository extends ObjectRepository
 {
     /**
-     * Returns the objects identifier.
+     * @param object $object
      *
-     * @return array
+     * @return mixed[]
      */
-    public function getObjectIdentifier($object)
+    public function getObjectIdentifier($object) : array
     {
         return $this->objectManager
             ->getClassMetadata(get_class($object))
@@ -35,15 +22,15 @@ class BasicObjectRepository extends ObjectRepository
     }
 
     /**
-     * Returns the identifier.
+     * @param mixed[] $data
      *
-     * @return array
+     * @return mixed[]
      */
-    public function getObjectIdentifierFromData(array $data)
+    public function getObjectIdentifierFromData(array $data) : array
     {
-        $identifier = array();
+        $identifier = [];
 
-        foreach ($this->class->identifier as $name) {
+        foreach ($this->class->getIdentifier() as $name) {
             $identifier[$name] = $data[$name];
         }
 
@@ -53,8 +40,8 @@ class BasicObjectRepository extends ObjectRepository
     /**
      * @param object $object
      */
-    public function merge($object)
+    public function merge($object) : void
     {
-        throw new \BadMethodCallException('Not implemented.');
+        throw new BadMethodCallException('Not implemented.');
     }
 }
