@@ -1,117 +1,71 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\ObjectRepository;
 
 use Doctrine\Common\EventManager;
 use Doctrine\SkeletonMapper\DataRepository\ObjectDataRepositoryInterface;
+use Doctrine\SkeletonMapper\Hydrator\ObjectHydratorInterface;
+use Doctrine\SkeletonMapper\Mapping\ClassMetadataInterface;
 use Doctrine\SkeletonMapper\ObjectFactory;
 use Doctrine\SkeletonMapper\ObjectManagerInterface;
-use Doctrine\SkeletonMapper\Hydrator\ObjectHydratorInterface;
+use InvalidArgumentException;
 
 /**
  * Base class for object repositories to extend from.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 abstract class ObjectRepository implements ObjectRepositoryInterface
 {
-    /**
-     * @var \Doctrine\SkeletonMapper\ObjectManagerInterface
-     */
+    /** @var ObjectManagerInterface */
     protected $objectManager;
 
-    /**
-     * @var \Doctrine\SkeletonMapper\DataRepository\ObjectDataRepositoryInterface
-     */
+    /** @var ObjectDataRepositoryInterface */
     protected $objectDataRepository;
 
-    /**
-     * @var \Doctrine\SkeletonMapper\ObjectFactory
-     */
+    /** @var ObjectFactory */
     protected $objectFactory;
 
-    /**
-     * @var \Doctrine\SkeletonMapper\Hydrator\ObjectHydratorInterface
-     */
+    /** @var ObjectHydratorInterface */
     protected $objectHydrator;
 
-    /**
-     * @var \Doctrine\Common\EventManager
-     */
+    /** @var EventManager */
     protected $eventManager;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $className;
 
-    /**
-     * @var \Doctrine\SkeletonMapper\Mapping\ClassMetadataInterface
-     */
+    /** @var ClassMetadataInterface */
     protected $class;
 
-    /**
-     * @param \Doctrine\SkeletonMapper\ObjectManagerInterface                       $objectManager
-     * @param \Doctrine\SkeletonMapper\DataRepository\ObjectDataRepositoryInterface $objectDataRepository
-     * @param \Doctrine\SkeletonMapper\ObjectFactory                                $objectFactory
-     * @param \Doctrine\SkeletonMapper\Hydrator\ObjectHydratorInterface             $objectHydrator
-     * @param \Doctrine\Common\EventManager                                         $eventManager
-     * @param string                                                                $className
-     */
     public function __construct(
         ObjectManagerInterface $objectManager,
         ObjectDataRepositoryInterface $objectDataRepository,
         ObjectFactory $objectFactory,
         ObjectHydratorInterface $objectHydrator,
         EventManager $eventManager,
-        $className = null)
-    {
-        $this->objectManager = $objectManager;
+        string $className
+    ) {
+        $this->objectManager        = $objectManager;
         $this->objectDataRepository = $objectDataRepository;
-        $this->objectFactory = $objectFactory;
-        $this->objectHydrator = $objectHydrator;
-        $this->eventManager = $eventManager;
-
-        if ($className !== null) {
-            $this->setClassName($className);
-        }
+        $this->objectFactory        = $objectFactory;
+        $this->objectHydrator       = $objectHydrator;
+        $this->eventManager         = $eventManager;
+        $this->setClassName($className);
     }
 
     /**
      * Returns the class name of the object managed by the repository.
-     *
-     * @return string
      */
-    public function getClassName()
+    public function getClassName() : string
     {
         return $this->className;
     }
 
-    /**
-     * @param string $className
-     */
-    public function setClassName($className)
+    public function setClassName(string $className) : void
     {
         $this->className = $className;
-        $this->class = $this->objectManager->getClassMetadata($this->className);
+        $this->class     = $this->objectManager->getClassMetadata($this->className);
     }
 
     /**
@@ -119,7 +73,7 @@ abstract class ObjectRepository implements ObjectRepositoryInterface
      *
      * @param mixed $id The identifier.
      *
-     * @return object The object.
+     * @return object|null The object.
      */
     public function find($id)
     {
@@ -128,59 +82,57 @@ abstract class ObjectRepository implements ObjectRepositoryInterface
         );
     }
 
-    /**
-     * Finds all objects in the repository.
-     *
-     * @return array The objects.
-     */
-    public function findAll()
+   /**
+    * Finds all objects in the repository.
+    *
+    * @return object[] The objects.
+    */
+    public function findAll() : array
     {
         $objectsData = $this->objectDataRepository->findAll();
 
-        $objects = array();
+        $objects = [];
         foreach ($objectsData as $objectData) {
-            $objects[] = $this->getOrCreateObject($objectData);
+            $object = $this->getOrCreateObject($objectData);
+
+            if ($object === null) {
+                throw new InvalidArgumentException('Could not create object.');
+            }
+
+            $objects[] = $object;
         }
 
         return $objects;
     }
 
     /**
-     * Finds objects by a set of criteria.
-     *
-     * Optionally sorting and limiting details can be passed. An implementation may throw
-     * an UnexpectedValueException if certain values of the sorting or limiting details are
-     * not supported.
-     *
-     * @param array      $criteria
-     * @param array|null $orderBy
-     * @param int|null   $limit
-     * @param int|null   $offset
-     *
-     * @return array The objects.
-     *
-     * @throws \UnexpectedValueException
+     * {@inheritDoc}
      */
-    public function findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+    public function findBy(array $criteria, ?array $orderBy = null, $limit = null, $offset = null) : array
     {
         $objectsData = $this->objectDataRepository->findBy(
-            $criteria, $orderBy, $limit, $offset
+            $criteria,
+            $orderBy,
+            $limit,
+            $offset
         );
 
-        $objects = array();
+        $objects = [];
         foreach ($objectsData as $objectData) {
-            $objects[] = $this->getOrCreateObject($objectData);
+            $object = $this->getOrCreateObject($objectData);
+
+            if ($object === null) {
+                throw new InvalidArgumentException('Could not create object.');
+            }
+
+            $objects[] = $object;
         }
 
         return $objects;
     }
 
     /**
-     * Finds a single object by a set of criteria.
-     *
-     * @param array $criteria The criteria.
-     *
-     * @return object The object.
+     * {@inheritDoc}
      */
     public function findOneBy(array $criteria)
     {
@@ -192,46 +144,50 @@ abstract class ObjectRepository implements ObjectRepositoryInterface
     /**
      * @param object $object
      */
-    public function refresh($object)
+    public function refresh($object) : void
     {
         $data = $this->objectDataRepository
             ->find($this->getObjectIdentifier($object));
+
+        if ($data === null) {
+            throw new InvalidArgumentException('Could not find object to refresh.');
+        }
 
         $this->hydrate($object, $data);
     }
 
     /**
-     * @param object $object
-     * @param array  $data
+     * @param object  $object
+     * @param mixed[] $data
      */
-    public function hydrate($object, array $data)
+    public function hydrate($object, array $data) : void
     {
         $this->objectHydrator->hydrate($object, $data);
     }
 
     /**
-     * @param string $className
      *
      * @return object
      */
-    public function create($className)
+    public function create(string $className)
     {
         return $this->objectFactory->create($className);
     }
 
     /**
-     * @param array|null $data
+     * @param mixed[] $data
      *
      * @return object|null
      */
-    protected function getOrCreateObject(array $data = null)
+    protected function getOrCreateObject(?array $data = null)
     {
         if ($data === null) {
-            return;
+            return null;
         }
 
         return $this->objectManager->getOrCreateObject(
-            $this->getClassName(), $data
+            $this->getClassName(),
+            $data
         );
     }
 }
