@@ -1,54 +1,27 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\Persister;
 
+use function sprintf;
+
 /**
  * Class responsible for retrieving ObjectPersister instances.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 class ObjectPersisterFactory implements ObjectPersisterFactoryInterface
 {
-    /**
-     * @var array
-     */
-    private $persisters = array();
+    /** @var ObjectPersisterInterface[] */
+    private $persisters = [];
 
-    /**
-     * @param string                                                $className
-     * @param \Doctrine\Common\Persistence\ObjectPersisterInterface $objectPersister
-     */
-    public function addObjectPersister($className, ObjectPersisterInterface $objectPersister)
+    public function addObjectPersister(string $className, ObjectPersisterInterface $objectPersister) : void
     {
         $this->persisters[$className] = $objectPersister;
     }
 
-    /**
-     * @param string $className
-     *
-     * @return \Doctrine\Common\Persistence\ObjectPersisterInterface
-     */
-    public function getPersister($className)
+    public function getPersister(string $className) : ObjectPersisterInterface
     {
-        if (!isset($this->persisters[$className])) {
+        if (! isset($this->persisters[$className])) {
             throw new \InvalidArgumentException(sprintf('ObjectPersister with class name %s was not found', $className));
         }
 
@@ -56,9 +29,9 @@ class ObjectPersisterFactory implements ObjectPersisterFactoryInterface
     }
 
     /**
-     * @return array
+     * @return ObjectPersisterInterface[]
      */
-    public function getPersisters()
+    public function getPersisters() : array
     {
         return $this->persisters;
     }
