@@ -1,44 +1,24 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\UnitOfWork;
 
 class ChangeSet
 {
-    /**
-     * @var object
-     */
+    /** @var object */
     private $object;
 
-    /**
-     * @var array
-     */
-    private $changes = array();
+    /** @var Change[] */
+    private $changes = [];
 
     /**
-     * @param object $object
-     * @param array  $changes
+     * @param object   $object
+     * @param Change[] $changes
      */
-    public function __construct($object, array $changes = array())
+    public function __construct($object, array $changes = [])
     {
-        $this->object = $object;
+        $this->object  = $object;
         $this->changes = $changes;
     }
 
@@ -50,39 +30,26 @@ class ChangeSet
         return $this->object;
     }
 
-    /**
-     * @param \Doctrine\SkeletonMapper\UnitOfWork\Change $change
-     */
-    public function addChange(Change $change)
+    public function addChange(Change $change) : void
     {
         $this->changes[$change->getPropertyName()] = $change;
     }
 
     /**
-     * @return array
+     * @return Change[]
      */
-    public function getChanges()
+    public function getChanges() : array
     {
         return $this->changes;
     }
 
-    /**
-     * @param string $fieldName
-     *
-     * @return bool
-     */
-    public function hasChangedField($fieldName)
+    public function hasChangedField(string $fieldName) : bool
     {
         return isset($this->changes[$fieldName]);
     }
 
-    /**
-     * @param string $fieldName
-     *
-     * @return \Doctrine\SkeletonMapper\UnitOfWork\Change
-     */
-    public function getFieldChange($fieldName)
+    public function getFieldChange(string $fieldName) : ?Change
     {
-        return isset($this->changes[$fieldName]) ? $this->changes[$fieldName] : null;
+        return $this->changes[$fieldName] ?? null;
     }
 }
