@@ -1,51 +1,32 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\UnitOfWork;
 
+use function spl_object_hash;
+
 class ChangeSets
 {
-    /**
-     * @var array
-     */
-    private $changeSets = array();
+    /** @var ChangeSet[] */
+    private $changeSets = [];
 
     /**
-     * @param object                                     $object
-     * @param \Doctrine\SkeletonMapper\UnitOfWork\Change $change
+     * @param object $object
      */
-    public function addObjectChange($object, Change $change)
+    public function addObjectChange($object, Change $change) : void
     {
         $this->getObjectChangeSet($object)->addChange($change);
     }
 
     /**
      * @param object $object
-     *
-     * @return \Doctrine\SkeletonMapper\UnitOfWork\ChangeSet
      */
-    public function getObjectChangeSet($object)
+    public function getObjectChangeSet($object) : ChangeSet
     {
         $oid = spl_object_hash($object);
 
-        if (!isset($this->changeSets[$oid])) {
+        if (! isset($this->changeSets[$oid])) {
             $this->changeSets[$oid] = new ChangeSet($object);
         }
 
