@@ -1,21 +1,23 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Doctrine\SkeletonMapper\Tests\Collections;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\SkeletonMapper\Collections\LazyCollection;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class LazyCollectionTest extends PHPUnit_Framework_TestCase
+class LazyCollectionTest extends TestCase
 {
-    public function testLoad()
+    public function testLoad() : void
     {
         $wrappedCollection = new ArrayCollection();
 
-        $collection = new LazyCollection(function() use ($wrappedCollection) {
+        $collection = new LazyCollection(function () use ($wrappedCollection) {
             return $wrappedCollection;
         });
 
-        $this->assertSame($wrappedCollection, $collection->getCollection());
+        self::assertSame($wrappedCollection, $collection->getCollection());
     }
 }
