@@ -1,117 +1,128 @@
 <?php
 
-namespace Doctrine\SkeletonMapper\Tests\Functional;
+declare(strict_types=1);
+
+namespace Doctrine\SkeletonMapper\Tests\Event;
 
 use Doctrine\SkeletonMapper\Event\PreUpdateEventArgs;
+use Doctrine\SkeletonMapper\ObjectManagerInterface;
 use Doctrine\SkeletonMapper\UnitOfWork\Change;
-use PHPUnit_Framework_TestCase;
+use Doctrine\SkeletonMapper\UnitOfWork\ChangeSet;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
+use stdClass;
 
-class PreUpdateEventArgsTest extends PHPUnit_Framework_TestCase
+class PreUpdateEventArgsTest extends TestCase
 {
+    /** @var ObjectManagerInterface */
     private $objectManager;
+
+    /** @var stdClass */
+    private $object;
+
+    /** @var ChangeSet|MockObject */
+    private $changeSet;
+
+    /** @var PreUpdateEventArgs */
     private $event;
 
-    protected function setUp()
+    public function testGetObjectChangeSet() : void
     {
-        $this->objectManager = $this->getMockBuilder('Doctrine\SkeletonMapper\ObjectManagerInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->object = new \stdClass();
-
-        $this->changeSet = $this->getMockBuilder('Doctrine\SkeletonMapper\UnitOfWork\ChangeSet')
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $this->event = new PreUpdateEventArgs($this->object, $this->objectManager, $this->changeSet);
+        self::assertSame($this->changeSet, $this->event->getObjectChangeSet());
     }
 
-    public function testGetObjectChangeSet()
+    public function testHasChangedField() : void
     {
-        $this->assertSame($this->changeSet, $this->event->getObjectChangeSet());
-    }
-
-    public function testHasChangedField()
-    {
-        $this->changeSet->expects($this->once())
+        $this->changeSet->expects(self::once())
             ->method('hasChangedField')
             ->with('username')
-            ->will($this->returnValue(true));
+            ->will(self::returnValue(true));
 
-        $this->assertTrue($this->event->hasChangedField('username'));
+        self::assertTrue($this->event->hasChangedField('username'));
     }
 
-    public function testGetOldValue()
+    public function testGetOldValue() : void
     {
         $change = new Change('username', 'jwage', 'jonwage');
 
-        $this->changeSet->expects($this->once())
+        $this->changeSet->expects(self::once())
             ->method('getFieldChange')
             ->with('username')
-            ->will($this->returnValue($change));
+            ->will(self::returnValue($change));
 
-        $this->assertEquals('jwage', $this->event->getOldValue('username'));
+        self::assertEquals('jwage', $this->event->getOldValue('username'));
     }
 
-    public function testGetOldValueReturnsNull()
+    public function testGetOldValueReturnsNull() : void
     {
-        $this->changeSet->expects($this->once())
+        $this->changeSet->expects(self::once())
             ->method('getFieldChange')
             ->with('username')
-            ->will($this->returnValue(null));
+            ->will(self::returnValue(null));
 
-        $this->assertNull($this->event->getOldValue('username'));
+        self::assertNull($this->event->getOldValue('username'));
     }
 
-    public function testGetNewValue()
-    {
-        $change = new Change('username', 'jwage', 'jonwage');
-
-        $this->changeSet->expects($this->once())
-            ->method('getFieldChange')
-            ->with('username')
-            ->will($this->returnValue($change));
-
-        $this->assertEquals('jonwage', $this->event->getNewValue('username'));
-    }
-
-    public function testGetNewValueReturnsNull()
-    {
-        $this->changeSet->expects($this->once())
-            ->method('getFieldChange')
-            ->with('username')
-            ->will($this->returnValue(null));
-
-        $this->assertNull($this->event->getNewValue('username'));
-    }
-
-    public function testSetNewValue()
+    public function testGetNewValue() : void
     {
         $change = new Change('username', 'jwage', 'jonwage');
 
-        $this->changeSet->expects($this->any())
+        $this->changeSet->expects(self::once())
             ->method('getFieldChange')
             ->with('username')
-            ->will($this->returnValue($change));
+            ->will(self::returnValue($change));
+
+        self::assertEquals('jonwage', $this->event->getNewValue('username'));
+    }
+
+    public function testGetNewValueReturnsNull() : void
+    {
+        $this->changeSet->expects(self::once())
+            ->method('getFieldChange')
+            ->with('username')
+            ->will(self::returnValue(null));
+
+        self::assertNull($this->event->getNewValue('username'));
+    }
+
+    public function testSetNewValue() : void
+    {
+        $change = new Change('username', 'jwage', 'jonwage');
+
+        $this->changeSet->expects(self::any())
+            ->method('getFieldChange')
+            ->with('username')
+            ->will(self::returnValue($change));
 
         $this->event->setNewValue('username', 'jonathan');
 
-        $this->assertEquals('jonathan', $this->event->getNewValue('username'));
+        self::assertEquals('jonathan', $this->event->getNewValue('username'));
     }
 
-    public function testSetNewValueForUnchangedField()
+    public function testSetNewValueForUnchangedField() : void
     {
         $change = new Change('username', null, 'jonwage');
 
-        $this->changeSet->expects($this->once())
+        $this->changeSet->expects(self::once())
             ->method('getFieldChange')
             ->with('username')
-            ->will($this->returnValue(null));
+            ->will(self::returnValue(null));
 
-        $this->changeSet->expects($this->once())
+        $this->changeSet->expects(self::once())
             ->method('addChange')
             ->with($change);
 
         $this->event->setNewValue('username', 'jonwage');
+    }
+
+    protected function setUp() : void
+    {
+        $this->objectManager = $this->createMock(ObjectManagerInterface::class);
+
+        $this->object = new stdClass();
+
+        $this->changeSet = $this->createMock(ChangeSet::class);
+
+        $this->event = new PreUpdateEventArgs($this->object, $this->objectManager, $this->changeSet);
     }
 }
