@@ -1,27 +1,30 @@
 <?php
 
-namespace Doctrine\SkeletonMapper\Tests\Persister;
+declare(strict_types=1);
+
+namespace Doctrine\SkeletonMapper\Tests;
 
 use Doctrine\SkeletonMapper\ObjectFactory;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
+use stdClass;
 
 /**
  * @group unit
  */
-class ObjectFactoryTest extends PHPUnit_Framework_TestCase
+class ObjectFactoryTest extends TestCase
 {
-    public function testCreate()
+    public function testCreate() : void
     {
         $factory = new ObjectFactoryStub();
-        $object = $factory->create('stdClass');
-        $this->assertInstanceOf('stdClass', $object);
+        $object  = $factory->create(stdClass::class);
+        self::assertInstanceOf(stdClass::class, $object);
     }
 
-    public function testCreateWithReflectionMethod()
+    public function testCreateWithReflectionMethod() : void
     {
         $factory = new ObjectFactoryReflectionMethodStub();
-        $object = $factory->create('stdClass');
-        $this->assertInstanceOf('stdClass', $object);
+        $object  = $factory->create(stdClass::class);
+        self::assertInstanceOf(stdClass::class, $object);
     }
 }
 
@@ -31,7 +34,7 @@ class ObjectFactoryStub extends ObjectFactory
 
 class ObjectFactoryReflectionMethodStub extends ObjectFactory
 {
-    protected function isReflectionMethodAvailable()
+    protected function isReflectionMethodAvailable() : bool
     {
         return true;
     }
