@@ -1,58 +1,93 @@
 <?php
 
-namespace Doctrine\SkeletonMapper\Tests\Functional;
+declare(strict_types=1);
 
-use Doctrine\SkeletonMapper\ObjectRepository\BasicObjectRepository;
+namespace Doctrine\SkeletonMapper\Tests\ObjectRepository;
+
+use BadMethodCallException;
+use Doctrine\Common\EventManager;
+use Doctrine\SkeletonMapper\DataRepository\ObjectDataRepositoryInterface;
+use Doctrine\SkeletonMapper\Hydrator\ObjectHydratorInterface;
 use Doctrine\SkeletonMapper\Mapping\ClassMetadata;
-use PHPUnit_Framework_TestCase;
+use Doctrine\SkeletonMapper\ObjectFactory;
+use Doctrine\SkeletonMapper\ObjectManagerInterface;
+use Doctrine\SkeletonMapper\ObjectRepository\BasicObjectRepository;
+use PHPUnit\Framework\MockObject\MockObject;
+use PHPUnit\Framework\TestCase;
 
 /**
  * @group unit
  */
-class BasicObjectRepositoryTest extends PHPUnit_Framework_TestCase
+class BasicObjectRepositoryTest extends TestCase
 {
+    /** @var ObjectManagerInterface|MockObject */
     private $objectManager;
+
+    /** @var ObjectDataRepositoryInterface|MockObject */
     private $objectDataRepository;
+
+    /** @var ObjectFactory|MockObject */
     private $objectFactory;
+
+    /** @var ObjectHydratorInterface|MockObject */
     private $hydrator;
+
+    /** @var EventManager|MockObject */
     private $eventManager;
+
+    /** @var ClassMetadata */
     private $classMetadata;
+
+    /** @var BasicObjectRepository */
     private $repository;
-    private $testClassName = 'Doctrine\SkeletonMapper\Tests\Functional\BasicObjectRepositoryTestModel';
 
-    protected function setUp()
+    /** @var string */
+    private $testClassName = BasicObjectRepositoryTestModel::class;
+
+    public function testGetObjectIdentifier() : void
     {
-        $this->objectManager = $this->getMockBuilder('Doctrine\SkeletonMapper\ObjectManagerInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $object     = new BasicObjectRepositoryTestModel();
+        $object->id = 1;
 
-        $this->objectDataRepository = $this->getMockBuilder('Doctrine\SkeletonMapper\DataRepository\ObjectDataRepositoryInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $data = ['id' => 1];
+        self::assertEquals($data, $this->repository->getObjectIdentifier($object));
+    }
 
-        $this->objectFactory = $this->getMockBuilder('Doctrine\SkeletonMapper\ObjectFactory')
-            ->disableOriginalConstructor()
-            ->getMock();
+    public function testGetObjectIdentifierFromData() : void
+    {
+        $data = ['id' => 1];
+        self::assertEquals($data, $this->repository->getObjectIdentifierFromData($data));
+    }
 
-        $this->hydrator = $this->getMockBuilder('Doctrine\SkeletonMapper\Hydrator\ObjectHydratorInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
+    public function testMerge() : void
+    {
+        $this->expectException(BadMethodCallException::class);
+        $this->expectExceptionMessage('Not implemented.');
 
-        $this->eventManager = $this->getMockBuilder('Doctrine\Common\EventManager')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $this->repository->merge(new \stdClass());
+    }
 
-        $this->classMetadata = new ClassMetadata($this->testClassName);
-        $this->classMetadata->identifier = array('id');
-        $this->classMetadata->identifierFieldNames = array('id');
-        $this->classMetadata->mapField(array(
-            'fieldName' => 'id',
-        ));
+    protected function setUp() : void
+    {
+        $this->objectManager = $this->createMock(ObjectManagerInterface::class);
 
-        $this->objectManager->expects($this->any())
+        $this->objectDataRepository = $this->createMock(ObjectDataRepositoryInterface::class);
+
+        $this->objectFactory = $this->createMock(ObjectFactory::class);
+
+        $this->hydrator = $this->createMock(ObjectHydratorInterface::class);
+
+        $this->eventManager = $this->createMock(EventManager::class);
+
+        $this->classMetadata                       = new ClassMetadata($this->testClassName);
+        $this->classMetadata->identifier           = ['id'];
+        $this->classMetadata->identifierFieldNames = ['id'];
+        $this->classMetadata->mapField(['fieldName' => 'id']);
+
+        $this->objectManager->expects(self::any())
             ->method('getClassMetadata')
             ->with($this->testClassName)
-            ->will($this->returnValue($this->classMetadata));
+            ->will(self::returnValue($this->classMetadata));
 
         $this->repository = new BasicObjectRepository(
             $this->objectManager,
@@ -63,33 +98,10 @@ class BasicObjectRepositoryTest extends PHPUnit_Framework_TestCase
             $this->testClassName
         );
     }
-
-    public function testGetObjectIdentifier()
-    {
-        $object = new BasicObjectRepositoryTestModel();
-        $object->id = 1;
-
-        $data = array('id' => 1);
-        $this->assertEquals($data, $this->repository->getObjectIdentifier($object));
-    }
-
-    public function testGetObjectIdentifierFromData()
-    {
-        $data = array('id' => 1);
-        $this->assertEquals($data, $this->repository->getObjectIdentifierFromData($data));
-    }
-
-    /**
-     * @expectedException BadMethodCallException
-     * @expectedExceptionMessage Not implemented.
-     */
-    public function testMerge()
-    {
-        $this->repository->merge(new \stdClass());
-    }
 }
 
 class BasicObjectRepositoryTestModel
 {
+    /** @var int */
     public $id;
 }
