@@ -1,29 +1,13 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\DataRepository;
 
+use UnexpectedValueException;
+
 /**
  * Interface that object data repositories must implement.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 interface ObjectDataRepositoryInterface
 {
@@ -32,16 +16,16 @@ interface ObjectDataRepositoryInterface
      *
      * @param mixed $id The identifier.
      *
-     * @return array The objects array of data.
+     * @return mixed[] The objects array of data.
      */
-    public function find($id);
+    public function find($id) : ?array;
 
     /**
      * Finds all object data in the repository.
      *
-     * @return array The objects data.
+     * @return mixed[][] The objects data.
      */
-    public function findAll();
+    public function findAll() : array;
 
     /**
      * Finds objects data by a set of criteria.
@@ -50,23 +34,26 @@ interface ObjectDataRepositoryInterface
      * an UnexpectedValueException if certain values of the sorting or limiting details are
      * not supported.
      *
-     * @param array      $criteria
-     * @param array|null $orderBy
-     * @param int|null   $limit
-     * @param int|null   $offset
+     * @param mixed[]      $criteria
+     * @param mixed[]|null $orderBy
      *
-     * @return array The objects.
+     * @return mixed[][] The objects data.
      *
-     * @throws \UnexpectedValueException
+     * @throws UnexpectedValueException
      */
-    public function findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null);
+    public function findBy(
+        array $criteria,
+        ?array $orderBy = null,
+        ?int $limit = null,
+        ?int $offset = null
+    ) : array;
 
     /**
      * Finds a single objects data by a set of criteria.
      *
-     * @param array $criteria The criteria.
+     * @param mixed[] $criteria The criteria.
      *
-     * @return array The objects array of data
+     * @return mixed[] The objects array of data
      */
-    public function findOneBy(array $criteria);
+    public function findOneBy(array $criteria) : ?array;
 }
