@@ -1,22 +1,6 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\Event;
 
@@ -24,8 +8,6 @@ use Doctrine\Common\Persistence\Event\OnClearEventArgs as BaseOnClearEventArgs;
 
 /**
  * Provides event arguments for the onClear event.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 class OnClearEventArgs extends BaseOnClearEventArgs
 {
@@ -33,9 +15,8 @@ class OnClearEventArgs extends BaseOnClearEventArgs
      * Returns the name of the object class that is cleared, or null if all
      * are cleared.
      *
-     * @return string|null
      */
-    public function getObjectClass()
+    public function getObjectClass() : ?string
     {
         return $this->getEntityClass();
     }
@@ -43,9 +24,8 @@ class OnClearEventArgs extends BaseOnClearEventArgs
     /**
      * Returns whether this event clears all objects.
      *
-     * @return bool
      */
-    public function clearsAllObjects()
+    public function clearsAllObjects() : bool
     {
         return $this->clearsAllEntities();
     }
