@@ -1,22 +1,6 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\Event;
 
@@ -26,28 +10,22 @@ use Doctrine\SkeletonMapper\UnitOfWork\ChangeSet;
 
 /**
  * Class that holds event arguments for a preUpdate event.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 class PreUpdateEventArgs extends LifecycleEventArgs
 {
-    /**
-     * @var \Doctrine\SkeletonMapper\UnitOfWork\ChangeSet
-     */
+    /** @var ChangeSet */
     private $objectChangeSet;
 
     /**
-     * Constructor.
      *
-     * @param object                                          $object
-     * @param \Doctrine\SkeletonMapper\ObjectManagerInterface $objectManager
-     * @param \Doctrine\SkeletonMapper\UnitOfWork\ChangeSet   $changeSet
+     *
+     * @param object $object
      */
     public function __construct(
         $object,
         ObjectManagerInterface $objectManager,
-        ChangeSet $changeSet)
-    {
+        ChangeSet $changeSet
+    ) {
         parent::__construct($object, $objectManager);
         $this->objectChangeSet = $changeSet;
     }
@@ -55,9 +33,8 @@ class PreUpdateEventArgs extends LifecycleEventArgs
     /**
      * Retrieves the object changeset.
      *
-     * @return \Doctrine\SkeletonMapper\UnitOfWork\ChangeSet
      */
-    public function getObjectChangeSet()
+    public function getObjectChangeSet() : ChangeSet
     {
         return $this->objectChangeSet;
     }
@@ -65,11 +42,9 @@ class PreUpdateEventArgs extends LifecycleEventArgs
     /**
      * Checks if field has a changeset.
      *
-     * @param string $field
      *
-     * @return bool
      */
-    public function hasChangedField($field)
+    public function hasChangedField(string $field) : bool
     {
         return $this->objectChangeSet->hasChangedField($field);
     }
@@ -77,13 +52,14 @@ class PreUpdateEventArgs extends LifecycleEventArgs
     /**
      * Gets the old value of the changeset of the changed field.
      *
-     * @param string $field
      *
      * @return mixed
      */
-    public function getOldValue($field)
+    public function getOldValue(string $field)
     {
-        if ($change = $this->objectChangeSet->getFieldChange($field)) {
+        $change = $this->objectChangeSet->getFieldChange($field);
+
+        if ($change !== null) {
             return $change->getOldValue();
         }
     }
@@ -91,13 +67,14 @@ class PreUpdateEventArgs extends LifecycleEventArgs
     /**
      * Gets the new value of the changeset of the changed field.
      *
-     * @param string $field
      *
      * @return mixed
      */
-    public function getNewValue($field)
+    public function getNewValue(string $field)
     {
-        if ($change = $this->objectChangeSet->getFieldChange($field)) {
+        $change = $this->objectChangeSet->getFieldChange($field);
+
+        if ($change !== null) {
             return $change->getNewValue();
         }
     }
@@ -105,12 +82,13 @@ class PreUpdateEventArgs extends LifecycleEventArgs
     /**
      * Sets the new value of this field.
      *
-     * @param string $field
-     * @param mixed  $value
+     * @param mixed $value
      */
-    public function setNewValue($field, $value)
+    public function setNewValue(string $field, $value) : void
     {
-        if ($change = $this->objectChangeSet->getFieldChange($field)) {
+        $change = $this->objectChangeSet->getFieldChange($field);
+
+        if ($change !== null) {
             $change->setNewValue($value);
         } else {
             $this->objectChangeSet->addChange(new Change($field, null, $value));
