@@ -1,22 +1,6 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\Mapping;
 
@@ -24,9 +8,36 @@ use Doctrine\Common\Persistence\Mapping\ClassMetadata as BaseClassMetadata;
 
 /**
  * Interface for class metadata instances.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 interface ClassMetadataInterface extends BaseClassMetadata
 {
+    /**
+     * @param mixed[] $identifier
+     */
+    public function setIdentifier(array $identifier) : void;
+
+    /**
+     * @param string[] $identifierFieldNames
+     */
+    public function setIdentifierFieldNames(array $identifierFieldNames) : void;
+
+    /**
+     * @param mixed[] $mapping
+     */
+    public function mapField(array $mapping) : void;
+
+    /**
+     * @return mixed[][]
+     */
+    public function getFieldMappings() : array;
+
+    public function hasLifecycleCallbacks(string $eventName) : bool;
+
+    /**
+     * @param object       $object
+     * @param null|mixed[] $arguments
+     */
+    public function invokeLifecycleCallbacks(string $event, $object, ?array $arguments = null) : void;
+
+    public function addLifecycleCallback(string $callback, string $event) : void;
 }
