@@ -1,49 +1,41 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper;
 
 use Doctrine\Common\Persistence\ObjectManager as BaseObjectManagerInterface;
+use Doctrine\SkeletonMapper\Mapping\ClassMetadataInterface;
+use Doctrine\SkeletonMapper\ObjectRepository\ObjectRepositoryInterface;
 
 /**
  * Interface that object managers must implement.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 interface ObjectManagerInterface extends BaseObjectManagerInterface
 {
     /**
      * @param object $object
      */
-    public function update($object);
+    public function update($object) : void;
 
     /**
-     * @param string $className
-     * @param array $data
+     * @param mixed[] $data
      *
      * @return object
      */
-    public function getOrCreateObject($className, array $data);
+    public function getOrCreateObject(string $className, array $data);
+
+    public function getUnitOfWork() : UnitOfWork;
 
     /**
-     * @return UnitOfWork
+     * @param string $className
+     *
+     * @return ObjectRepositoryInterface
      */
-    public function getUnitOfWork();
+    public function getRepository($className);
+
+    /**
+     * @param string $className
+     */
+    public function getClassMetadata($className) : ClassMetadataInterface;
 }
