@@ -1,22 +1,6 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\Persister;
 
@@ -24,8 +8,6 @@ use Doctrine\SkeletonMapper\UnitOfWork\ChangeSet;
 
 /**
  * Interface that object persisters must implement.
- *
- * @author Jonathan H. Wage <jonwage@gmail.com>
  */
 interface ObjectPersisterInterface
 {
@@ -34,51 +16,49 @@ interface ObjectPersisterInterface
      *
      * @param object $object
      *
-     * @return array
+     * @return mixed[]
      */
-    public function preparePersistChangeSet($object);
+    public function preparePersistChangeSet($object) : array;
 
     /**
      * Prepares an object update changeset for update.
      *
-     * @param object                                        $object
-     * @param \Doctrine\SkeletonMapper\UnitOfWork\ChangeSet $changeSet
+     * @param object $object
      *
-     * @return array
+     * @return mixed[]
      */
-    public function prepareUpdateChangeSet($object, ChangeSet $changeSet);
+    public function prepareUpdateChangeSet($object, ChangeSet $changeSet) : array;
 
     /**
      * Performs operation to write object to the database.
      *
      * @param object $object
      *
-     * @return array $objectData
+     * @return mixed[] $objectData
      */
-    public function persistObject($object);
+    public function persistObject($object) : array;
 
     /**
      * Assign identifier to object.
      *
-     * @param object $object
-     * @param array  $identifier
+     * @param object  $object
+     * @param mixed[] $identifier
      */
-    public function assignIdentifier($object, array $identifier);
+    public function assignIdentifier($object, array $identifier) : void;
 
     /**
      * Performs operation to update object in the database.
      *
-     * @param object                                        $object
-     * @param \Doctrine\SkeletonMapper\UnitOfWork\ChangeSet $changeSet
+     * @param object $object
      *
-     * @return array $objectData
+     * @return mixed[] $objectData
      */
-    public function updateObject($object, ChangeSet $changeSet);
+    public function updateObject($object, ChangeSet $changeSet) : array;
 
     /**
      * Performs operation to remove object in the database.
      *
      * @param object $object
      */
-    public function removeObject($object);
+    public function removeObject($object) : void;
 }
