@@ -1,58 +1,32 @@
 <?php
 
-/*
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- * This software consists of voluntary contributions made by many individuals
- * and is licensed under the MIT license. For more information, see
- * <http://www.doctrine-project.org>.
- */
+declare(strict_types=1);
 
 namespace Doctrine\SkeletonMapper\UnitOfWork;
 
 class Change
 {
-    /**
-     * @var string
-     */
+    /** @var string */
     private $propertyName;
 
-    /**
-     * @var mixed
-     */
+    /** @var mixed */
     private $oldValue;
 
-    /**
-     * @var mixed
-     */
+    /** @var mixed */
     private $newValue;
 
     /**
-     * @param string $propertyName
-     * @param mixed  $oldValue
-     * @param mixed  $newValue
+     * @param mixed $oldValue
+     * @param mixed $newValue
      */
-    public function __construct($propertyName, $oldValue, $newValue)
+    public function __construct(string $propertyName, $oldValue, $newValue)
     {
         $this->propertyName = $propertyName;
-        $this->oldValue = $oldValue;
-        $this->newValue = $newValue;
+        $this->oldValue     = $oldValue;
+        $this->newValue     = $newValue;
     }
 
-    /**
-     * @return string
-     */
-    public function getPropertyName()
+    public function getPropertyName() : string
     {
         return $this->propertyName;
     }
@@ -76,7 +50,7 @@ class Change
     /**
      * @param mixed $newValue
      */
-    public function setNewValue($newValue)
+    public function setNewValue($newValue) : void
     {
         $this->newValue = $newValue;
     }
