@@ -1,10 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Doctrine\SkeletonMapper\Tests;
 
 interface DataTesterInterface
 {
-    public function find($id);
-    public function set($id, $key, $value);
-    public function count();
+    /**
+     * @return mixed[]
+     */
+    public function find(int $id) : ?array;
+
+    /**
+     * @param mixed $value
+     */
+    public function set(int $id, string $key, $value) : void;
+
+    public function count() : int;
 }
